+++ common/src/main/java/it/crystalnest/prometheus/mixin/CampfireBlockMixin.java
@@ -36,7 +36,7 @@ public abstract class CampfireBlockMixin implements FireTypeChanger {
 
   @Override
   public void setFireType(ResourceLocation fireType) {
-    this.fireType = fireType;
+    this.fireType = FireManager.ensure(fireType);
   }
 
   /**